feat: move animations to hook
