+++ apps/mobile/src/components/atoms/Animated/index.ts
@@ -1,0 +1,1 @@
+export * from './useAnimatedHeader';
